requirement.txt
discord.py
pandas
matplotlib
python-dotenv
